+++ HISTORY.md
@@ -26,6 +26,7 @@
     * new "get pen attribute" reporter
     * new "write" command in pen category (used to be "label" in tools)
     * new "map","keep", "combine" and "for each" primitives in list category
+    * new "for" loop primitive in the Control category
     * added "neg", "lg" (log2) and "2^" selectors to monadic function reporter in Operators
     * added "^" reporter (power of) in the Operators category
     * added "width" and "height" as attribute selectors of the OF primitive for the stage
@@ -69,8 +70,9 @@
     * French
 
 ### 2019-04-24
-* Threads: new "combine" primitive in list category
+* Threads, Objects: new "combine" primitive in list category
 * Threads: added type-assertions for the new HOF prims
+* Threads, Objects: new "for" loop primitive in Control category
 
 ### 2019-04-23
 * Threads: fixed JS stack overflow issue for MAP primitive